requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Django==2.1.14
django-storages==1.7.2
django-user-agents==0.4.0
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-core==1.0.3
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
idna==2.8
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyMySQL==0.9.3
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
ua-parser==0.8.0
urllib3==1.25.7
user-agents==2.0
